+++ changelog.md
@@ -5,13 +5,20 @@ Alle belangrijke wijzigingen in **Beherenbestanden.ps1** worden hieronder bijgeh
 ---
 
 ## [4.8.1] - In uitvoering
-### Bugfix
+### Bugfixes
 - Icoontje van snelkoppeling op het bureaublad verwees naar het verkeerde bestand. 
   Opgelost door bij de start, eenmalig, opnieuw een snelkoppeling te maken.
+- Powershell 7 werd niet geïnstalleerd als je dit bij **Instellingen** wijzigt.
+  Nu wordt het programma opnieuw opgestart en PowerShell 7 geïnstalleerd.
+- Bij sommige Windows 11 pc's kreeg je een foutmelding als je de console wilde sluiten.
+  Opgelost door een bepaalde module niet meer te gebruiken.
 
 ### Overige wijzigingen
 - Script om snelkoppeling te maken aangepast zodat dit zonder gebruikersinvoer
   kan worden uitgevoerd, kan werken met PowerShell 7 en een startmap kan meegeven.
+- Readme is aangepast. Aantal teksten verduidelijkt.
+
+---
 
 ## [4.8.0] - 2026-06-19
 ### Nieuwe functies